Split config controller into a package
The 1449-line controllers/config.py mixed seven concerns: dashboard
settings, the metadata-sidecar transaction layer, user preferences,
remote-build settings, the label store, esptool chip detection, and the
ConfigController itself. Replace it with a controllers/config/ package
following the canonical split shape (devices/, firmware/, remote_build/):
one submodule per concern plus __init__.py re-exporting the full public
surface (and the private symbols the test suite reaches into) so every
existing 'from ...controllers.config import X' caller keeps working.

Behavior-free restructure. Largest new module is 392 lines.

Test mock targets that patched the old single module's seams
(get_serial_ports, run_subprocess_capture, resolve_storage_path,
_run_esptool, _METADATA_LOCK, stat) follow those seams into their new
submodules; same fakes, same assertions.
